+++ source/internal_extensions/omni.cuopt.visualization/omni/cuopt/visualization/generate_waypoint_graph.py
@@ -7,8 +7,6 @@
 # disclosure or distribution of this material and related documentation
 # without an express license agreement from NVIDIA CORPORATION or
 # its affiliates is strictly prohibited.
-import json
-
 from omni.kit.material.library import CreateAndBindMdlMaterialFromLibrary
 from pxr import Gf, Sdf, UsdGeom, UsdShade